requirements.txt
anthropic>=0.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
anthropic
beautifulsoup4
python-dotenv
requests
